requirements.txt
opencv-python
pytesseract
gtts
ultralytics
python-telegram-bot==13.8
